+++ polyjuice/polyjuice.py
@@ -3,12 +3,15 @@ docstr = """
 Polyjuice
 
 Usage:
-    polyjuice.py [-h] (<input_path> <output_path>) [-z <zip_output_path>]
-    polyjuice.py [-h] [-z <zip_output_path>]
+    polyjuice.py (-h | --help)
+    polyjuice.py [-l | --log] (<input_path> <output_path>) [-z <zip_output_path>]
+    polyjuice.py [-l | --log] [-z <zip_output_path>]
 
 Options:
-  -h --help                                     show this message and exit
+  -h --help                                     Show this message and exit
   -z --zip                                      Archives the output folder
+  -l --log                                      Give progress of program
+
 Instructions:
     Run polyjuice on the ISO file or on the Extracted DICOM folder. This will give an ouput folder
 containing dicom files with unneccessary tags removed
@@ -17,17 +20,19 @@ $ ./polyjuice.py path_to_ISOfile.iso path_to_OutputFolder
 
 Inorder to ZIP your Cleaned Output Directory
 $ ./polyjuice.py -z path_to_ISOfile.iso path_to_OutputFolder Path_to_Zipped_file
+
 """
 import os
 import os.path
 import shutil
 import yaml
 from docopt import docopt
-from dicom_image import DicomImage
+from filch import DicomCaretaker
 
 INPUT_DIR = '<input_path>'
 OUTPUT_DIR = '<output_path>'
 ZIP_DIR = '<zip_output_path>'
+_print_log = '--log'
 
 def consult_book(out_dir):
     if not os.path.exists(out_dir):
@@ -45,38 +50,34 @@ def raid_snapes_cupboard():
                 print(exc)
         return (config)
 
-def brew_potion(dicom_file, in_dir, out_dir, removals, alterations):
+def brew_potion(dicom_file, in_dir, out_dir, deletions, modifications, verbose):
     for path, subdirs, files in os.walk(in_dir):
         for name in files:
             print os.path.join(path, name)
             try:
                 with open(os.path.join(path, name)) as working_file:
-                    print("Working on {}".format(name))
-                    dataset = dicom_file.scrub(working_file, removals, alterations)
+                    if verbose:
+                        print("Working on {}".format(name))
+                    dataset = dicom_file.scrub(working_file, deletions, modifications, verbose)
                     dicom_file.save_output(dataset, out_dir, name)
             except Exception, e:
                 print("{} failed".format(name))
                 print (str(e))
 
 def main(args):
-    import sys
-    if not (args.get(INPUT_DIR) and args.get(OUTPUT_DIR)):
-        print("Please Enter Input and Output files ")
-        sys.exit()
-    # dicom_dir, out_dir = sys.argv[1:]
     dicom_dir = args[INPUT_DIR]
     out_dir = args[OUTPUT_DIR]
     consult_book(out_dir)
 
-    dicom_file = DicomImage()
+    dicom_file = DicomCaretaker()
 
     in_dir = dicom_file.start(dicom_dir)
 
     config = raid_snapes_cupboard()
-    removals = config.get('deletions')
-    alterations = config.get('modifications')
+    deletions = config.get('deletions')
+    modifications = config.get('modifications')
 
-    #brew_potion(dicom_file, in_dir, out_dir, removals, alterations)
+    brew_potion(dicom_file, in_dir, out_dir, deletions, modifications, args[_print_log])
 
     # Working on converting into ZIP folder
     if(args.get(ZIP_DIR)):